LPD-77476 Organize Formik components into a dedicated directory
