+++ modules/apps/export-import/export-import-web/src/main/resources/META-INF/resources/revamp/js/components/forms/formik/FormikFieldText.tsx
@@ -1,0 +1,19 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+import React from 'react';
+
+import FieldText, {FieldTextProps} from '../FieldText';
+import {FormikWrapper} from './FormikWrapper';
+
+export function FormikFieldText(props: FieldTextProps) {
+	return (
+		<FormikWrapper name={props.name} required={props.required}>
+			{({errorMessage, field}) => (
+				<FieldText {...props} {...field} errorMessage={errorMessage} />
+			)}
+		</FormikWrapper>
+	);
+}
