+++ modules/apps/export-import/export-import-web/src/main/resources/META-INF/resources/revamp/js/pages/export/steps/SetupStep.tsx
@@ -10,7 +10,7 @@ import React from 'react';
 import {
 	FormikFieldMultiCheckbox,
 	FormikFieldText,
-} from '../../../components/forms/FormikFields';
+} from '../../../components/forms/formik';
 import {mockPorletDataHandlerSections} from '../../../utils/mockPorletDataHandlerSections';
 
 export default function SetupStep() {
